Tooling refactor
- Removed Babel, and moved tests as well as the main source to CommonJS,
  greatly simplyify tooling,
- The above resulted in an easier-to-bundle code for browsers as a side-effect,
- Removed esdoc, as it wasn't working anyway.
- Added the engines field in line with the travis tests.
- Replaced esdoc with jsdoc (As the former was not working anyways)
