--- doc.json
@@ -1,8 +1,0 @@
-{
-  "source": "./src",
-  "destination": "./docs",
-  "test": {
-    "type": "mocha",
-    "source": "./test"
-  }
-}
