+++ package.json
@@ -2,19 +2,21 @@
   "name": "json-truncate",
   "version": "0.0.0-semantically-released",
   "description": "A way to truncate a json object.",
-  "main": "index.js",
+  "main": "json-truncate.js",
   "scripts": {
     "precommit": "npm run lint && npm run coverage",
     "commit": "git-cz",
     "lint": "standard",
     "coverage": "nyc npm t && npm run coverage-report",
     "coverage-report": "nyc report --reporter=lcov",
-    "doc": "esdoc -c doc.json",
-    "test": "npm run build && _mocha --compilers js:babel-register",
+    "test": "mocha",
     "prebuild": "rimraf dist",
+    "doc": "jsdoc json-truncate.js -d docs",
     "predoc": "rimraf docs",
-    "build": "babel --copy-files --out-dir dist src",
     "travis": "npm run precommit"
+  },
+  "engines": {
+    "node": ">=6.0.0"
   },
   "repository": {
     "type": "git",
@@ -35,47 +37,32 @@
     "js"
   ],
   "author": "Matt Steele <matt@omnionline.us> (http://omnionline.us)",
+  "contributors": [
+    "Viktor Hubert <rpgmorpheus@gmail.com> (https://hubertviktor.com)"
+  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/mrsteele/json-truncate/issues"
   },
   "homepage": "https://github.com/mrsteele/json-truncate#readme",
   "devDependencies": {
-    "babel-cli": "^6.10.1",
-    "babel-eslint": "^10.0.2",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-register": "^6.9.0",
     "chai": "^4.0.2",
     "commitizen": "^4.0.0",
     "cz-conventional-changelog": "^3.0.0",
-    "esdoc": "^1.0.1",
+    "jsdoc": "^3.6.3",
     "mocha": "^5.0.0",
     "nyc": "^14.1.0",
     "rimraf": "^3.0.0",
     "standard": "^13.0.1"
   },
-  "files": [
-    "dist",
-    "src",
-    "README.md"
-  ],
   "config": {
     "commitizen": {
       "path": "./node_modules/cz-conventional-changelog"
     }
   },
-  "standard": {
-    "parser": "babel-eslint",
-    "ignore": "/dist/"
-  },
-  "babel": {
-    "presets": [
-      "es2015"
-    ]
-  },
   "nyc": {
     "include": [
-      "src/**/*.js"
+      "json-truncate.js"
     ]
   }
 }
